--- src/plugin.ts
@@ -1,102 +1,0 @@
-import path from 'node:path';
-
-import * as docgen from 'react-docgen-typescript';
-// import { matcher } from 'micromatch';
-import ts from 'typescript';
-import type * as webpack from 'webpack';
-
-interface TypescriptOptions {
-  /**
-   * Specify the location of the tsconfig.json to use. Can not be used with
-   * compilerOptions.
-   **/
-  tsconfigPath?: string;
-  /** Specify TypeScript compiler options. Can not be used with tsconfigPath. */
-  compilerOptions?: ts.CompilerOptions;
-}
-
-export type PluginOptions = TypescriptOptions & {
-  /** Glob patterns to ignore */
-  exclude?: string[];
-  /** Glob patterns to include. defaults to ts|tsx */
-  include?: string[];
-};
-
-/** Get the contents of the tsconfig in the system */
-function getTSConfigFile(tsconfigPath: string): ts.ParsedCommandLine {
-  try {
-    const basePath = path.dirname(tsconfigPath);
-    const configFile = ts.readConfigFile(tsconfigPath, ts.sys.readFile);
-
-    return ts.parseJsonConfigFileContent(configFile.config, ts.sys, basePath, {}, tsconfigPath);
-  } catch {
-    return {} as ts.ParsedCommandLine;
-  }
-}
-
-// /** Create a glob matching function. */
-// const matchGlob = (globs?: string[]) => {
-//   const matchers = (globs || []).map((g) => matcher(g, { dot: true }));
-
-//   return (filename: string) => Boolean(filename && matchers.find((match) => match(filename)));
-// };
-
-/** Inject typescript docgen information into modules at the end of a build */
-export class DocgenPlugin implements webpack.WebpackPluginInstance {
-  public static defaultOptions = {
-    setDisplayName: true,
-    typePropName: 'type',
-    docgenCollectionName: 'STORYBOOK_REACT_CLASSES',
-  };
-
-  private name = 'React Docgen Typescript Plugin';
-  private options: PluginOptions;
-
-  constructor(options: PluginOptions = {}) {
-    this.options = options;
-  }
-
-  apply(compiler: webpack.Compiler): void {
-    const options = this.getOptions();
-    const pluginName = 'DocGenPlugin';
-    compiler.hooks.compilation.tap(pluginName, (_, { normalModuleFactory }) => {
-      normalModuleFactory.hooks.afterResolve.tap(pluginName, (result) => {
-        if (/\.(tsx?)$/.test(result.request)) {
-          result.createData.loaders = result.createData.loaders || [];
-          result.createData.loaders?.push({
-            loader: require.resolve(__dirname + '/../dist/loader.js'),
-            options: { parser: docgen.withCompilerOptions(options.compilerOptions, { shouldIncludeExpression: true }) },
-          });
-        }
-      });
-    });
-  }
-
-  getOptions(): {
-    compilerOptions: ts.CompilerOptions;
-  } {
-    const { tsconfigPath = './tsconfig.json', compilerOptions: userCompilerOptions } = this.options;
-
-    let compilerOptions = {
-      jsx: ts.JsxEmit.React,
-      module: ts.ModuleKind.CommonJS,
-      target: ts.ScriptTarget.Latest,
-    };
-
-    if (userCompilerOptions) {
-      compilerOptions = {
-        ...compilerOptions,
-        ...userCompilerOptions,
-      };
-    } else {
-      const { options: tsOptions } = getTSConfigFile(tsconfigPath);
-      compilerOptions = { ...compilerOptions, ...tsOptions };
-    }
-
-    return {
-      compilerOptions,
-    };
-  }
-}
-
-export type DocgenPluginType = typeof DocgenPlugin;
