+++ vitest.config.ts
@@ -1,8 +1,8 @@
-import { defineConfig } from "vitest/config";
+import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    environment: "node",
-    include: ["src/**/*.{test,spec}.{js,ts,tsx}"],
+    environment: 'node',
+    include: ['src/**/*.{test,spec}.{js,ts,tsx}'],
   },
 });
